+++ sdk/resourcemanager/operationalinsights/armoperationalinsights/managementgroups_client.go
@@ -19,7 +19,7 @@ import (
 // ManagementGroupsClient contains the methods for the ManagementGroups group.
 // Don't use this type directly, use NewManagementGroupsClient() instead.
 //
-// Generated from API version 2025-07-01
+// Generated from API version 2026-03-01
 type ManagementGroupsClient struct {
 	internal       *arm.Client
 	subscriptionID string
@@ -89,7 +89,7 @@ func (client *ManagementGroupsClient) listCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", version20250701)
+	reqQP.Set("api-version", version20260301)
 	req.Raw().URL.RawQuery = strings.ReplaceAll(reqQP.Encode(), "+", "%20")
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil